+++ MainBackend/Services/Interfaces/IBarInventoryService.cs
@@ -5,6 +5,6 @@ namespace MainBackend.Services.Interfaces;
 public interface IBarInventoryService
 {
     Task<ICollection<BarInventory>> GetBarItems();
-    Task<bool> AddBarItem(string name, int quantity, float price);
+    Task<bool> AddBarItem(string name, int quantity, decimal price);
     Task<bool> DeleteBarItem(int id);
 }
